fdts: stm32mp151-prtt1a.dts: Add support for eMMC on PRTT1C variant
While the PRTT1A technically does not have an eMMC chip, the pins required
are not connected, so this patch does no harm. OTOH, having different
TF-A binaries for every board is unnecessarily cumbersome, so we try to
make one for all three. The HW has been designed with this in mind.

Signed-off-by: David Jander <david@protonic.nl>