plat: st: stm32cubeprogrammer_uart.c: Add support for command 0x32U
This requires a modified stm32prog tool, but enables using command 0x32 to
download bigger packets. This is useful in combination with higher baud
rates (1.5MBaud for example) to drastically improve throughput.
The only difference with command 0x31 (regular download command) is the
packet length value is changed from U8 to U16-BE.

Signed-off-by: David Jander <david@protonic.nl>